feat: move MongoDB related files to db/ folder
